anki: improve responses to QXs questions by including 'All'
e.g. if the answer to "How should you play 7Xo (where 7 is higher)?" is "Fold:
72o+" then that means fold all combinations. Now this is included in the answer
to make it easier to remember.

Also make a small improvement to test_anki.py so it also detects when no
"question" could be found in a note. This makes it easier to debug some
potential issues in the future.