feat: -t/--tag can be specified multiple times (breaking change)
The purpose is to make the subcommands more similar to each other.

Previously it was possible to specify multiple tags by separating them
with white spaces, but now -t/--tag must be included every time. This
would otherwise create a conflict with the unlimited number of
positional arguments of the 'hand' subcommand.

Example:

    # Previously
    anki-poker-master --tags poker learning 2024 -o package.apkg -s scenario.yml

    # Now
    anki-poker-master range --tag poker --tag learning -t 2024 -o package.apkg -s scenario.yml