fix: add a unique field for hand history Anki cards to avoid duplicates
By default Anki performs a duplicate detection based on the first field
of the note type. For cases such as this one where some fields are very
similar to each other (e.g. the "Title") and the uniqueness derives from
the combination of multiple fields, that is not helpful.

An alternative to this change would be to change the order of the field
and use one of the questions as the first field. That has the following
downsides:

* Two notes might have the same "question" in a spot but the context
  being different require a different answer
* Questions and answers are the least likely to be edited in Anki since
  they contain complex HTML so they should not be displayed first